fixed _all_ bugs of 0.7.3
git-svn-id: https://svn.code.sf.net/p/mars-game/code@214 249382d3-5f29-4b26-bf2c-91a77bf04246

+++ include/SpaceObjects/Ship.hpp
@@ -42,7 +42,7 @@ class Ship: public MobileSpaceObject {
         void onShockWave(Player* damageSource, float intensity);
 
         void setDamageSource(Player* evilOne);
-        void drainLife(Player* source, int amount, Vector2f const& direction, float waitForOtherDamage = 0.001f);
+        void drainLife(Player* source, float amount, Vector2f const& direction, float waitForOtherDamage = 0.001f);
 
         void heal(Player* source, int amount);
         void refuel(Player* source, int amount);
